♻️ Move ConnectorOptions and DefaultConnectorOptions to opts
